+++ src/Components/Icon/Icon.stories.tsx
@@ -1,10 +1,10 @@
 // Libraries
-import React, {CSSProperties} from 'react'
+import React, {CSSProperties, RefObject, createRef} from 'react'
 import marked from 'marked'
 
 // Storybook
 import {storiesOf} from '@storybook/react'
-import {withKnobs, select, number, object} from '@storybook/addon-knobs'
+import {withKnobs, select, object} from '@storybook/addon-knobs'
 import {mapEnumKeys} from '../../Utils/storybook'
 import {jsxDecorator} from 'storybook-addon-jsx'
 
@@ -21,21 +21,32 @@ const iconStories = storiesOf('Atomic|Icon', module)
   .addDecorator(withKnobs)
   .addDecorator(jsxDecorator)
 
-const iconStyleExample: CSSProperties = {color: '#6BDFFF'}
+const iconStyleExample: CSSProperties = {color: '#6BDFFF', fontSize: '60px'}
 
 iconStories.add(
   'Example',
-  () => (
-    <div
-      className="story--example"
-      style={{fontSize: `${number('Font Size', 60)}px`}}
-    >
-      <Icon
-        glyph={IconFont[select('glyph', mapEnumKeys(IconFont), 'Bell')]}
-        style={object('style', iconStyleExample)}
-      />
-    </div>
-  ),
+  () => {
+    const iconRef: RefObject<HTMLSpanElement> = createRef()
+
+    const logIconRef = (): void => {
+      /* eslint-disable */
+      console.log(iconRef.current)
+      /* eslint-enable */
+    }
+
+    return (
+      <div className="story--example">
+        <Icon
+          ref={iconRef}
+          glyph={IconFont[select('glyph', mapEnumKeys(IconFont), 'Bell')]}
+          style={object('style', iconStyleExample)}
+        />
+        <div className="story--test-buttons">
+          <button onClick={logIconRef}>Log Ref</button>
+        </div>
+      </div>
+    )
+  },
   {
     readme: {
       content: marked(IconReadme),