+++ src/Components/Label/Label.tsx
@@ -1,15 +1,17 @@
 // Libraries
-import React, {Component, CSSProperties, MouseEvent} from 'react'
-import chroma from 'chroma-js'
+import React, {FunctionComponent, MouseEvent, useState, forwardRef} from 'react'
 import classnames from 'classnames'
 
 // Types
-import {StandardClassProps, ComponentSize, InfluxColors} from '../../Types'
+import {StandardFunctionProps, ComponentSize, InfluxColors} from '../../Types'
+
+// Utils
+import {generateLabelStyle} from '../../Utils'
 
 // Styles
 import './Label.scss'
 
-interface Props extends StandardClassProps {
+export interface LabelProps extends StandardFunctionProps {
   /** Unique value to be returned when Label is clicked */
   id: string
   /** Name of the Label, appears inside the label */
@@ -23,162 +25,121 @@ interface Props extends StandardClassProps {
   /** Optional delete handler, if passed in the delete button is rendered */
   onDelete?: (id: string) => void
   /** Size of Label */
-  size: ComponentSize
+  size?: ComponentSize
 }
 
-interface State {
-  isMouseOver: boolean
-}
+export type LabelRef = HTMLDivElement
 
-export class Label extends Component<Props, State> {
-  public static readonly displayName = 'Label'
+export const Label = forwardRef<LabelRef, LabelProps>(
+  (
+    {
+      id,
+      name,
+      size = ComponentSize.ExtraSmall,
+      style,
+      color,
+      testID = 'label--pill',
+      onClick,
+      onDelete,
+      className,
+      description,
+    },
+    ref
+  ) => {
+    const [isMouseOver, setHoverState] = useState(false)
 
-  public static defaultProps = {
-    size: ComponentSize.ExtraSmall,
-    testID: 'label--pill',
-  }
-
-  constructor(props: Props) {
-    super(props)
-
-    this.state = {
-      isMouseOver: false,
-    }
-  }
-
-  public render() {
-    const {name, testID, id} = this.props
-
-    return (
-      <div
-        className={this.className}
-        onMouseEnter={this.handleMouseEnter}
-        onMouseLeave={this.handleMouseLeave}
-        style={this.style}
-        title={this.title}
-        id={id}
-      >
-        <span
-          className="cf-label--name"
-          onClick={this.handleClick}
-          data-testid={`${testID} ${name}`}
-        >
-          {name}
-        </span>
-        {this.deleteButton}
-      </div>
-    )
-  }
-
-  private handleClick = (e: MouseEvent<HTMLDivElement>): void => {
-    const {id, onClick} = this.props
-
-    if (onClick) {
-      e.stopPropagation()
-      e.preventDefault()
-      onClick(id)
-    }
-  }
-
-  private handleDelete = (): void => {
-    const {id, onDelete} = this.props
-
-    if (onDelete) {
-      onDelete(id)
-    }
-  }
-
-  private handleMouseEnter = (): void => {
-    const {onClick} = this.props
-
-    if (onClick) {
-      this.setState({isMouseOver: true})
-    }
-  }
-
-  private handleMouseLeave = (): void => {
-    const {onClick} = this.props
-
-    if (onClick) {
-      this.setState({isMouseOver: false})
-    }
-  }
-
-  private get className(): string {
-    const {size, onClick, onDelete, className} = this.props
-
-    return classnames('cf-label', {
+    const labelClass = classnames('cf-label', {
       [`${className}`]: className,
       [`cf-label--${size}`]: size,
       'cf-label--deletable': onDelete,
       'cf-label--clickable': onClick,
     })
-  }
 
-  private get title(): string {
-    const {onClick, name, description} = this.props
-
-    if (onClick) {
-      return `Click to see all resources with the "${name}" label`
+    const handleClick = (e: MouseEvent<HTMLDivElement>): void => {
+      if (onClick) {
+        e.stopPropagation()
+        e.preventDefault()
+        onClick(id)
+      }
     }
 
-    return `${description}`
-  }
+    const handleDelete = (): void => {
+      if (onDelete) {
+        onDelete(id)
+      }
+    }
 
-  private get deleteButton(): JSX.Element | undefined {
-    const {onDelete, name, testID} = this.props
+    const handleToggleHoverState = (): void => {
+      if (onClick) {
+        setHoverState(!isMouseOver)
+      }
+    }
 
-    if (onDelete) {
-      return (
-        <button
-          className="cf-label--delete"
-          onClick={this.handleDelete}
-          type="button"
-          title={`Remove label "${name}"`}
-          data-testid={`${testID}--delete ${name}`}
+    const labelStyle = generateLabelStyle(color, !!onClick, isMouseOver, style)
+
+    return (
+      <div
+        onMouseEnter={handleToggleHoverState}
+        onMouseLeave={handleToggleHoverState}
+        className={labelClass}
+        style={labelStyle}
+        title={description}
+        ref={ref}
+        id={id}
+      >
+        <span
+          className="cf-label--name"
+          onClick={handleClick}
+          data-testid={`${testID} ${name}`}
         >
-          <div
-            className="cf-label--delete-x"
-            style={{backgroundColor: this.textColor}}
+          {name}
+        </span>
+        {onDelete && (
+          <LabelDeleteButton
+            onDelete={handleDelete}
+            name={name}
+            testID={testID}
+            color={`${labelStyle.color}`}
           />
-          <div
-            className="cf-label--delete-x"
-            style={{backgroundColor: this.textColor}}
-          />
-        </button>
-      )
-    }
-
-    return
+        )}
+      </div>
+    )
   }
+)
 
-  private get style(): CSSProperties {
-    const {isMouseOver} = this.state
-    const {color, onClick, style} = this.props
-
-    let backgroundColor = color
-
-    if (isMouseOver && onClick) {
-      backgroundColor = `${chroma(color).brighten(1)}`
-    }
-
-    return {
-      backgroundColor: `${backgroundColor}`,
-      color: this.textColor,
-      ...style,
-    }
-  }
-
-  private get textColor(): string {
-    const {color} = this.props
-
-    const darkContrast = chroma.contrast(color, InfluxColors.Kevlar)
-    const lightContrast = chroma.contrast(color, InfluxColors.White)
-
-    if (darkContrast > lightContrast) {
-      return InfluxColors.Kevlar
-    } else {
-      return InfluxColors.White
-    }
-  }
+interface DeleteButtonProps {
+  /** Name of the Label, appears inside the label */
+  name: string
+  /** Optional delete handler, if passed in the delete button is rendered */
+  onDelete: () => void
+  /** ID for Integration Tests */
+  testID: string
+  /** Ensures the X is the same color as the text label */
+  color: string
 }
+
+const LabelDeleteButton: FunctionComponent<DeleteButtonProps> = ({
+  onDelete,
+  name,
+  testID,
+  color,
+}) => {
+  const labelDeleteStyle = {backgroundColor: color}
+
+  return (
+    <button
+      className="cf-label--delete"
+      onClick={onDelete}
+      type="button"
+      title={`Remove label "${name}"`}
+      data-testid={`${testID}--delete ${name}`}
+    >
+      <div className="cf-label--delete-x" style={labelDeleteStyle} />
+      <div className="cf-label--delete-x" style={labelDeleteStyle} />
+    </button>
+  )
+}
+
+LabelDeleteButton.displayName = 'LabelDeleteButton'
+
+Label.displayName = 'Label'