+++ src/Components/Spinners/SpinnerContainer.tsx
@@ -1,29 +1,41 @@
 // Libraries
-import React, {Component} from 'react'
+import React, {forwardRef} from 'react'
 import classnames from 'classnames'
 
 // Types
-import {RemoteDataState, StandardClassProps} from '../../Types'
+import {RemoteDataState, StandardFunctionProps} from '../../Types'
 
 // Styles
 import './SpinnerContainer.scss'
 
-interface Props extends StandardClassProps {
+export interface SpinnerContainerProps extends StandardFunctionProps {
   /** Loading state */
   loading: RemoteDataState
   /** Spinner component */
   spinnerComponent: JSX.Element
 }
 
-export class SpinnerContainer extends Component<Props> {
-  public static readonly displayName = 'SpinnerContainer'
+export type SpinnerContainerRef = HTMLDivElement
 
-  public static defaultProps = {
-    testID: 'spinner-container',
-  }
-
-  public render() {
-    const {loading, children, spinnerComponent, testID, id, style} = this.props
+export const SpinnerContainer = forwardRef<
+  SpinnerContainerRef,
+  SpinnerContainerProps
+>(
+  (
+    {
+      id,
+      style,
+      testID = 'spinner-container',
+      loading,
+      children,
+      className,
+      spinnerComponent,
+    },
+    ref
+  ) => {
+    const spinnerContainerClass = classnames('cf-spinner-container', {
+      [`${className}`]: className,
+    })
 
     if (
       loading === RemoteDataState.Loading ||
@@ -31,9 +43,10 @@ export class SpinnerContainer extends Component<Props> {
     ) {
       return (
         <div
-          className={this.className}
+          className={spinnerContainerClass}
           data-testid={testID}
           id={id}
+          ref={ref}
           style={style}
         >
           {spinnerComponent}
@@ -41,12 +54,8 @@ export class SpinnerContainer extends Component<Props> {
       )
     }
 
-    return children
+    return <>{children}</>
   }
+)
 
-  private get className(): string {
-    const {className} = this.props
-
-    return classnames('cf-spinner-container', {[`${className}`]: className})
-  }
-}
+SpinnerContainer.displayName = 'SpinnerContainer'