+++ src/Components/Spinners/WaitingText.tsx
@@ -1,43 +1,38 @@
 // Libraries
-import React, {Component} from 'react'
+import React, {forwardRef} from 'react'
 import classnames from 'classnames'
 
 // Styles
 import './WaitingText.scss'
 
 // Types
-import {StandardClassProps} from '../../Types'
+import {StandardFunctionProps} from '../../Types'
 
-interface Props extends StandardClassProps {
+export interface WaitingTextProps extends StandardFunctionProps {
   /** Text to be displayed */
   text: string
 }
 
-export class WaitingText extends Component<Props> {
-  public static readonly displayName = 'WaitingText'
+export type WaitingTextRef = HTMLDivElement
 
-  public static defaultProps = {
-    testID: 'waiting-text',
-  }
-
-  public render() {
-    const {text, testID, id, style} = this.props
+export const WaitingText = forwardRef<WaitingTextRef, WaitingTextProps>(
+  ({id, text, style, testID = 'waiting-text', className}, ref) => {
+    const waitingTextClass = classnames('cf-waiting-text', {
+      [`${className}`]: className,
+    })
 
     return (
       <div
-        className={this.className}
-        data-testid={testID}
         id={id}
+        ref={ref}
         style={style}
+        data-testid={testID}
+        className={waitingTextClass}
       >
         {text}
       </div>
     )
   }
+)
 
-  private get className(): string {
-    const {className} = this.props
-
-    return classnames('cf-waiting-text', {[`${className}`]: className})
-  }
-}
+WaitingText.displayName = 'WaitingText'